Phase out duplication of code which locates dotnet.exe.
